refactor: 옵션 목록 조회 Port 변경
